chore(core): enforce OSS-EE boundary with import-linter contract (Phase 3b)
Add the "OSS core may not import from EE" forbidden contract: pocketpaw
must never statically import pocketpaw_ee. The single ignore is the admin
CLI (pocketpaw.tools.cli), which administers an EE install directly.
Runtime EE access via the entry-point registry is dynamic and correctly
invisible to import-linter's static analysis.

Adding the contract surfaced that create_pocket_and_session — moved to
pocketpaw_ee.cloud.pockets.agent_context in the previous commit — made
agent_context import models.session, breaking the pre-existing "Pockets
— Beanie writes only from service.py" contract. Relocate the function to
pockets/service.py, the designated Beanie-write layer. lint-imports:
16 kept, 0 broken.
