+++ ee/pocketpaw_ee/cloud/pockets/service.py
@@ -1517,6 +1517,109 @@ async def agent_create(
     return _agent_view_dict(doc), str(doc.id), None
 
 
+async def create_pocket_and_session(
+    spec: dict[str, Any],
+    session_key: str,
+    user_id: str | None = None,
+    workspace_id: str | None = None,
+) -> str | None:
+    """Create a pocket + linked chat session in MongoDB. Returns the pocket
+    id, or ``None`` on failure.
+
+    Backs the core ``pocketpaw.pockets`` PocketWriter extension point:
+    ``pocketpaw.agents.loop`` calls this when a local-mode CreatePocketTool
+    emits a ``pocket_event: created``. Identity arrives as explicit args
+    (threaded from ``InboundMessage.metadata``) rather than per-stream
+    ContextVars, because the agent loop has no SSE request scope. When
+    ``user_id`` / ``workspace_id`` are missing it falls back to legacy
+    heuristics — ``user.active_workspace`` first, then first-owned
+    workspace, then any workspace — so self-hosted single-user deployments
+    (CLI, Telegram) without JWT auth still work.
+    """
+    try:
+        from datetime import UTC, datetime
+
+        from pocketpaw_ee.cloud.models.session import Session
+        from pocketpaw_ee.cloud.models.user import User
+        from pocketpaw_ee.cloud.models.workspace import Workspace
+
+        # ── User selection ──────────────────────────────────────────────
+        # Prefer an explicitly-threaded user id so agent-created pockets
+        # land under the caller, not whichever user comes first out of Mongo.
+        user = None
+        if user_id:
+            try:
+                user = await User.get(PydanticObjectId(user_id))
+            except Exception:  # noqa: BLE001
+                logger.warning("Invalid cloud_user_id %r; falling back to first user", user_id)
+                user = None
+        if not user:
+            user = await User.find_one()
+        if not user:
+            logger.warning("Cannot create pocket — no user in DB")
+            return None
+        user_id = str(user.id)
+
+        # ── Workspace selection ─────────────────────────────────────────
+        # Priority: explicit workspace_id → user.active_workspace → first
+        # owned workspace → any workspace.
+        workspace = None
+        target_ws = workspace_id or getattr(user, "active_workspace", None)
+        if target_ws:
+            try:
+                workspace = await Workspace.get(PydanticObjectId(target_ws))
+            except Exception:  # noqa: BLE001
+                workspace = None
+        if not workspace:
+            workspace = await Workspace.find_one(Workspace.owner == user_id)
+        if not workspace:
+            workspace = await Workspace.find_one()
+        if not workspace:
+            logger.warning("Cannot create pocket — no workspace in DB")
+            return None
+        workspace_id = str(workspace.id)
+
+        # Create the pocket through the standard service entry point.
+        meta = spec.get("metadata", {})
+        pocket = await create(
+            workspace_id,
+            user_id,
+            CreatePocketRequest(
+                name=spec.get("title") or spec.get("name") or "Untitled",
+                description=spec.get("description", ""),
+                type=meta.get("category", "custom"),
+                icon="sparkles",
+                color=meta.get("color", "#0A84FF"),
+                rippleSpec=spec,
+            ),
+        )
+        pocket_id = str(pocket["_id"])
+
+        # Link (find-or-create) the chat session to the new pocket.
+        safe_key = session_key.replace(":", "_") if session_key else ""
+        if safe_key:
+            existing = await Session.find_one(Session.sessionId == safe_key)
+            if existing:
+                existing.pocket = pocket_id
+                await existing.save()
+            else:
+                session = Session(
+                    sessionId=safe_key,
+                    workspace=workspace_id,
+                    owner=user_id,
+                    title=spec.get("title") or "New Chat",
+                    pocket=pocket_id,
+                    lastActivity=datetime.now(UTC),
+                )
+                await session.insert()
+
+        logger.info("Created pocket %s + session %s in MongoDB", pocket_id, safe_key)
+        return pocket_id
+    except Exception:
+        logger.warning("Failed to create pocket/session in MongoDB", exc_info=True)
+        return None
+
+
 __all__ = [
     "access_via_share_link",
     "add_agent",
@@ -1541,6 +1644,7 @@ __all__ = [
     "agent_view",
     "create",
     "create_from_ripple_spec",
+    "create_pocket_and_session",
     "delete",
     "generate_share_link",
     "get",
